lambdas/data-subscriber-query/requirements.txt
aws-lambda-powertools
requests>=2.22.0
python-dateutil